requirements.txt
loguru==0.5.0
selenium==3.141.0
pytest==5.4.2
allure-pytest==2.8.15
pytest-html==2.1.1
pyyaml==5.3.1
